Created a custom hook useStockData and modified the stockSelector.js file
